feat: ensure DeleteTransaction implements LoadTransactionById

+++ src/data/usecases/transaction/delete-transaction/db-delete-transaction.ts
@@ -1,8 +1,13 @@
-import { DeleteTransactionRepository } from './../../../protocols/transaction/delete-transaction-repository'
-import { DeleteTransaction } from './../../../../domain/usecases/transaction/delete-transaction'
+import { LoadTransactionByIdRepository, DeleteTransaction, DeleteTransactionRepository } from './db-delete-transaction-protocols'
+
 export class DbDeleteTransaction implements DeleteTransaction {
-  constructor (private readonly deleteTransactionRepository: DeleteTransactionRepository) {}
+  constructor (private readonly deleteTransactionRepository: DeleteTransactionRepository,
+    private readonly loadTransactionByIdRepository: LoadTransactionByIdRepository
+  ) {}
+
   async delete (id: string): Promise<boolean> {
+    const transaction = await this.loadTransactionByIdRepository.loadById(id)
+    if (!transaction) return false
     await this.deleteTransactionRepository.delete(id)
     return true
   }
